Rework: extract AvatarFactory and MediaRequestDataFactory from CoilMediaFetcher
